+++ src/atoms/atoms.js
@@ -3,6 +3,7 @@ import { atom } from "jotai";
 export const userAtom = atom("");
 
 export const databasesAtom = atom([]);
+export const documentsAtom = atom([]);
 
 export const currentDBIdAtom = atom("");
 export const currentDBNameAtom = atom("");
